site: add sitemap and robots control in front matter
The `sitemap_exclude` and `robots` parameters were added to control search engine behavior on individual pages. The `sitemap_exclude` is used to decide if a page should be included in the `sitemap.xml` while the `robots` parameter provides instructions to web crawlers for indexing the page or following its links.